backend/api/requirements.txt
Flask==3.1.1
Flask-Cors==6.0.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
SQLAlchemy==2.0.41
requests==2.32.4